fixed format

+++ test/Chainlink.t.sol
@@ -6,23 +6,23 @@ import {Chainlink} from "../src/Chainlink.sol";
 
 contract ChainlinkTest is Test {
     Chainlink public chainlink;
-    
+
     // Chainlink price feed address on Scroll Sepolia
     address public constant PRICE_FEED_ADDRESS = 0x87dce67002e66C17BC0d723Fe20D736b80CAaFda;
-    
+
     function setUp() public {
         // Fork Scroll Sepolia testnet
         uint256 forkId = vm.createFork("https://sepolia-rpc.scroll.io");
         vm.selectFork(forkId);
-        
+
         // Deploy the Chainlink contract
         chainlink = new Chainlink();
     }
-    
+
     function test_GetChainlinkDataFeedLatestAnswer() public view {
         // Test calling the getChainlinkDataFeedLatestAnswer function
-        int answer = chainlink.getChainlinkDataFeedLatestAnswer();
-        
+        int256 answer = chainlink.getChainlinkDataFeedLatestAnswer();
+
         // Verify we got a valid price (should be > 0 for most assets)
         assertTrue(answer > 0, "Should return a valid price from Chainlink feed");
         // console.log("Chainlink price feed answer:", answer);
